#5 - refactor: 로그인 관련 옮기기
로그인 관련한 로직들을 authController로 옮김
