+++ src/main/java/com/example/peep/controller/StudentController.java
@@ -1,6 +1,5 @@
 package com.example.peep.controller;
 
-import com.example.peep.dto.JwtTokenDto;
 import com.example.peep.dto.StudentDto;
 import com.example.peep.service.StudentService;
 import lombok.RequiredArgsConstructor;
@@ -21,28 +20,5 @@ public class StudentController {
     @PostMapping("/new")
     public void newStudent(@RequestBody StudentDto studentDto) {
         studentService.newStudent(studentDto);
-    }
-
-    //TODO: authController 를 만들어 분리하기
-
-    @PostMapping("/login")
-    public JwtTokenDto login( @RequestHeader("Device-Id") String deviceId,
-                           @RequestBody StudentDto studentDto) {
-         return studentService.login(deviceId, studentDto);
-    }
-
-    @PostMapping("/refresh")
-    public JwtTokenDto refreshToken(@RequestHeader("Device-Id") String deviceId,
-                                 @RequestBody JwtTokenDto jwtTokenDto,
-                                 @RequestParam("userId") String userId) {
-
-        return studentService.refreshToken(deviceId, userId, jwtTokenDto);
-    }
-
-    @PostMapping("/logout")
-    public String logout(@RequestBody JwtTokenDto jwtTokenDto) {
-
-        studentService.logout(jwtTokenDto);
-        return "Logout successful";
     }
 }
